processors/create_csv_sentences/requirements.txt
python-docx>=0.8.11
nltk>=3.8
pandas>=1.5.0
chardet>=5.0.0
